pyproject.toml
[project]
name = "neural-soup-currency-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4==4.14.2",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "colorama==0.4.6",
    "contourpy==1.3.2",
    "coverage==7.12.0",
    "curl-cffi==0.13.0",
    "cycler==0.12.1",
    "decorator==5.2.1",
    "edfio==0.4.11",
    "edflib-python==1.0.8",
    "filelock==3.19.1",
    "fonttools==4.60.1",
    "frozendict==2.4.7",
    "fsspec==2025.9.0",
    "h5py==3.15.1",
    "idna==3.11",
    "joblib==1.5.2",
    "kiwisolver==1.4.9",
    "lazy-loader==0.4",
    "matplotlib==3.10.7",
    "memory-profiler==0.61.0",
    "mne==1.10.2",
    "mne-bids==0.17.0",
    "moabb==1.4.0",
    "mpmath==1.3.0",
    "multitasking==0.0.12",
    "networkx==3.3",
    "numpy==2.2.6",
    "packaging==25.0",
    "pandas==2.3.3",
    "peewee==3.18.3",
    "pillow==12.0.0",
    "platformdirs==4.5.0",
    "pooch==1.8.2",
    "protobuf==6.33.1",
    "psutil==7.1.3",
    "pycparser==2.23",
    "pyparsing==3.2.5",
    "pyriemann==0.9",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "scikit-learn==1.5.2",
    "scipy==1.15.3",
    "seaborn==0.12.2",
    "six==1.17.0",
    "soupsieve==2.8",
    "sympy==1.13.1",
    "threadpoolctl==3.6.0",
    "torch==2.6.0",
    "torchaudio==2.6.0",
    "torchvision==0.21.0",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "tzdata==2025.2",
    "urllib3==1.26.20",
    "websockets==15.0.1",
    "yfinance==0.2.66",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu" },
]
torchvision = [
  { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
